+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## [Unreleased]
 ### Fixed
-- Don't display all packages on code coverage view with checked 'Flatten Packages' 
+- Show Diff Coverage report notification if `diffCoverage` Gradle task failed due to violation rules
+- Don't display all packages on code coverage view with checked 'Flatten Packages'
 
 ## [0.0.4]
 ### Fixed